pyproject.toml
[project]
name = "vps-ssh-wrapper"
version = "0.1.0"
description = "Lightweight secure SSH client for VPS administration with persistent connections and CLI tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paramiko>=3.4.0",
    "python-dotenv>=1.0.0",
]
[project.scripts]
vps = "main:main"
